requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
bcrypt==4.1.3
blinker==1.8.1
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
datasets==2.19.1
dill==0.3.8
dnspython==2.6.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
faiss-cpu==1.8.0
fasttext-wheel==0.9.2
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.0
frozenlist==1.4.1
fsspec==2024.3.1
gensim==4.3.2
huggingface-hub==0.23.0
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.1.1
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
networkx==3.3
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
passlib==1.7.4
preshed==3.0.9
pyarrow==16.0.0
pyarrow-hotfix==0.6
pybind11==2.12.0
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.12.0
setuptools==69.5.1
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
tbb==2021.12.0
thinc==8.2.3
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.40.1
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.2
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
